Add camera snapshot support to skrl playback
